Guess That Number v3

+++ src/Util/handleAutoKick.js
@@ -1,8 +1,7 @@
-const handleDatabaseError = require('../util/handleDatabaseError.js');
+const handleDatabaseError = require('./handleDatabaseError');
 
-module.exports = {
-	interval: (1000 * 60 * 60),
-	execute: (bot, r) => {
+module.exports = (bot, r) => {
+	setInterval(() => {
 		r.table('games').run((error, response) => {
 			if (error) return handleDatabaseError(error);
 			if (response.length !== 0) {
@@ -11,16 +10,14 @@ module.exports = {
 						if (bot.users.get(index.id)) bot.users.get(index.id).getDMChannel().then((dm) => dm.createMessage('As of 12 hours ago, you started a ' + ((index.difficulty === 1) ? 'easy' : ((index.difficulty === 2) ? 'medium' : ((index.difficulty === 3) ? 'hard' : 'unknown'))) + ' difficulty game. To keep from getting too many active games, we have kicked you from it. If you would like to start another game, you may do so by running the `start` command again. Thank you.'));
 						r.table('games').get(index.id).delete().run((error) => {
 							if (error) return handleDatabaseError(error);
-							if (bot.toggle.has(index.id)) {
-								r.table('toggle').get(index.id).delete().run((error) => {
-									if (error) return handleDatabaseError(error);
-									bot.toggle.delete(index.id);
-								});
-							}
+							if (bot.toggle.has(index.id)) r.table('toggle').get(index.id).delete().run((error) => {
+								if (error) return handleDatabaseError(error);
+								bot.toggle.delete(index.id);
+							});
 						});
 					}
 				});
 			}
 		});
-	}
+	}, 1000 * 60 * 60);
 };
